Ghost click issue using Ionic
This issue is strictly the problem of this plugin, but without this change there's a serious issue for Ionic users which I don't think would have any workaround.

The windows version of the plugin basically just creates a number of `<select>` elements to look like a native picker. As discussed in this blog post: http://www.intertech.com/Blog/windows-phone-ionic-cordova-select-element-issue/ there are issues with double clicking on select elements in Ionic. The resolution is to just set the "data-tap-disabled" attribute on the elements. I've tested this out and it works for me. I can't see any other way Ionic users could get round this issue, hence my proposed (very small) changes. However I do appreciate this is slightly odd to put a framework-specific fix in here.